refactor: replace add note function
